Temporary Revert PKHeX.Core Back To 24.01.12 For Version 1.6.1

+++ PluginPile.Sorting/SortingPlugin.cs
@@ -25,7 +25,7 @@ public class SortingPlugin : PluginBase {
     SortByButton.Text = Language.MenuItemName;
     SortByButton.DropDownItems.Clear();
     ToolStripItemCollection sortItems = SortByButton.DropDownItems;
-    byte gen = SaveFileEditor.SAV.Generation;
+    int gen = SaveFileEditor.SAV.Generation;
     GameVersion version = SaveFileEditor.SAV.Version;
     bool isLetsGo = version == GameVersion.GP || version == GameVersion.GE;
     if (isLetsGo) {
